requirements.txt
lxml==4.6.4
urllib3==1.26.7
python-dotenv==0.17.1
